Finish both the SimulationResult and RestResult entities,repositories, and controllers

+++ rest/core/src/main/java/net/microfalx/heimdall/rest/core/overview/SimulationResultController.java
@@ -8,5 +8,5 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller("OverviewSimulationResultController")
 @DataSet(model = SimulationResult.class, timeFilter = false)
 @RequestMapping("/rest/simulation/result")
-public class SimulationResultController extends DataSetController<SimulationResult,Integer> {
+public class SimulationResultController extends DataSetController<SimulationResult,Long> {
 }
